Refactoring order_item_bloc and fixing another issues. Fixing #2
