+++ lib/features/orders/presentation/bloc/order_item_bloc.dart
@@ -1,31 +1,6 @@
-import 'package:equatable/equatable.dart';
+import 'package:eco_bites/features/orders/presentation/bloc/order_item_event.dart';
+import 'package:eco_bites/features/orders/presentation/bloc/order_item_state.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-
-class OrderItemState extends Equatable {
-
-  const OrderItemState({required this.status});
-  final String status;
-
-  @override
-  List<Object> get props => <Object>[status];
-}
-
-abstract class OrderItemEvent extends Equatable {
-  const OrderItemEvent();
-
-  @override
-  List<Object> get props => <Object>[];
-}
-
-class OrderStatusChanged extends OrderItemEvent {
-
-  const OrderStatusChanged(this.newStatus);
-  final String newStatus;
-
-  @override
-  List<Object> get props => <Object>[newStatus];
-}
-
 class OrderItemBloc extends Bloc<OrderItemEvent, OrderItemState> {
   OrderItemBloc(String initialStatus)
       : super(OrderItemState(status: initialStatus)) {
