+++ lib/features/orders/presentation/bloc/order_item_event.dart
@@ -1,0 +1,17 @@
+import 'package:equatable/equatable.dart';
+
+abstract class OrderItemEvent extends Equatable {
+  const OrderItemEvent();
+
+  @override
+  List<Object> get props => <Object>[];
+}
+
+class OrderStatusChanged extends OrderItemEvent {
+  const OrderStatusChanged(this.newStatus);
+
+  final String newStatus;
+
+  @override
+  List<Object> get props => <Object>[newStatus];
+}
