+++ lib/features/orders/presentation/bloc/order_item_state.dart
@@ -1,0 +1,10 @@
+import 'package:equatable/equatable.dart';
+
+class OrderItemState extends Equatable {
+  const OrderItemState({required this.status});
+
+  final String status;
+
+  @override
+  List<Object> get props => <Object>[status];
+}
